extra-handlers

See merge request seng440-2019/project2-team4!2

+++ app/src/main/java/me/nubuscu/hotpotato/model/dto/InGameUpdateMessage.kt
@@ -2,6 +2,6 @@ package me.nubuscu.hotpotato.model.dto
 
 
 data class InGameUpdateMessage(
-    val timeRemaining: Int,
+    val timeRemaining: Long,
     val dest: String //endpoint ID being sent to
 ) : Message("inGame")
